adding "accept" and "drop" chains
These chains are similar to "reject" chain.
Packets are processed inside those chains before they reach the
final "ACCEPT" or "DROP" target.

This can be useful in situations when you need to add a rule
to all packets which are accepted / dropped
(.e.g. to log all packets from a particular IP address)

Signed-off-by: Stepan Henek <stepan.henek@nic.cz>